- fixed menu print for active sub menus - don't permit to destroy post types post and page

+++ app/models/camaleon_cms/post_type.rb
@@ -159,6 +159,10 @@ class CamaleonCms::PostType < CamaleonCms::TermTaxonomy
 
   # destroy all custom field groups assigned to this post type
   def destroy_field_groups
+    if self.slug == "post" || self.slug == "page"
+      errors.add(:base, "This post type can not be deleted.")
+      return false
+    end
     self.get_field_groups.destroy_all
   end
